==> 01_zipcode.ipynb <==
# Lat Lon to ZIP Code for 2020 Crime Data
2023-09-29<br>
Evangeline Chang

# %pip install geopy

import pandas as pd
import numpy as np
from typing import Optional
from geopy.geocoders import Nominatim
from datetime import datetime
import time
import os

import warnings
warnings.filterwarnings("ignore", category=FutureWarning)

crime = pd.read_csv('data/Crime_Data_head20.csv', index_col=0)
print(crime.shape)
crime.head()

## Get Year
Seperate the dataset by years and save them respectively 

def get_year(in_file):
    for index, row in in_file.iterrows():
        date_str = row['DATE OCC']
        date_obj = datetime.strptime(date_str, "%m/%d/%Y %I:%M:%S %p")
        year = date_obj.year

        in_file.at[index, 'Year'] = int(year)

get_year(crime)

crime_2020 = crime.loc[crime['Year'] == 2020]

crime_2020.to_csv('data/crime_data_2020_head20.csv')

crime_2020_split = np.array_split(crime_2020, 10)

data_directory = 'data/Crime_split/'

if not os.path.exists(data_directory):
    os.makedirs(data_directory)

for i in range(0, 10):
    split_file_path = os.path.join(data_directory, f'crime_data_2020_split{i+1:02}_zip.csv')
    crime_2020_split[i].to_csv(split_file_path)

## Reverse Geocoding
Convert the latitude and longitude to ZIP codes

def latlong_zip(in_file, start_index=0, save_interval=1000, output_file=None):

    for index, row in in_file.iterrows():
        if index < start_index:
            continue
        
        try:
            lat = row['LAT']
            lon = row['LON']
            coordinates = f"{lat}, {lon}"
            geolocator = Nominatim(user_agent="geoapizipcodepractice")
            location = geolocator.reverse(coordinates)
            location_dict = location.raw
            zipcode = location_dict.get('address', {}).get('postcode', None)
            
            in_file.at[index, 'ZIP'] = zipcode

            if (index + 1) % 100 == 0:
                print(f"Processed {index + 1} rows. Lat is {lat}, long is {lon}, zipcode is {zipcode}. Sleeping for 1 second.")
                time.sleep(1)

            if (index + 1) % save_interval == 0:
                if output_file:
                    in_file.to_csv(output_file)
                    print(f'File saved till row {index + 1}')

        except Exception as e:
            print(str(e))

# first split the dataset, then run reverse geocoding
output_directory = 'data/Crime_zip/'

if not os.path.exists(output_directory):
    os.makedirs(output_directory)

for i in range(0, 10):
    print(f'crime_2020_split_{i+1:02}')
    output_file_path = os.path.join(output_directory, f'crime_data_2020_split{i+1:02}_zip.csv')
    latlong_zip(crime_2020_split[i], 0, output_file=output_file_path)
    crime_2020_split[i].to_csv(output_file_path)

# it should take less than 15 seconds to run 20 rows

combined_df = pd.concat(crime_2020_split)
combined_df.to_csv('data/crime_data_2020_zipcode_final.csv')
combined_df.head()
