==> 03_crime_rate.ipynb <==
# Calculate Crime Rate by ZIP codes
2023-10-05<br>
Evangeline Chang

import pandas as pd
import os

The code below is used to determine whether the crime is categorized as `personal` or `property`.

# crime = pd.read_csv('data/Crime_zip/crime_data_2020_zip_final.csv', index_col=0)

# personal = [110, 113, 121, 122, 
#             210, 220, 230, 231, 235, 236, 237, 250, 251, 
#             434, 435, 436, 437, 439, 
#             622, 623, 624, 625, 626, 627, 647, 651, 652, 653, 654, 661, 
#             753, 755, 756, 760, 761, 762, 763, 
#             805, 806, 810, 812, 813, 814, 815, 820, 821, 822, 845, 850, 860, 865, 870, 
#             900, 901, 902, 903, 906, 910, 920, 921, 922, 926, 928, 930, 932, 940, 942, 943, 944, 946, 954, 956]

# property = [310, 320, 330, 331, 341, 343, 345, 347, 349, 350, 351, 352, 353, 354, 
#             410, 420, 421, 433, 438, 440, 441, 442, 443, 444, 445, 446, 450, 
#             451, 452, 453, 470, 471, 472, 473, 474, 475, 480, 485, 487, 491, 
#             510, 520, 522, 648, 649, 660, 662, 664, 666, 668, 670, 
#             740, 745, 880, 886, 888, 890, 
#             931, 933, 949, 950, 951]

# def crime_type(crime_code):
#     if crime_code in personal:
#         return 'personal'
#     else:
#         return 'property'
    
# crime['Crime_Type'] = crime['Crm Cd'].apply(crime_type)
# crime_counts = crime.groupby(['ZIP', 'Crime_Type']).size().reset_index(name='Crime_Count')

Since I can't upload the whole `crime_data_2020_zip_final` dataset, you can use the `crime_count` file, which I created using the code above.

crime_counts = pd.read_csv('data/crime_count.csv', index_col=0)
crime_counts

popu = pd.read_csv('data/DECENNIALDHC2020.P1-Data.csv', skiprows=[1])
popu = popu.rename(columns={"P1_001N": "Population"})

def get_zip(in_file):
    for index, row in in_file.iterrows():
        zip_original = row['NAME']
        zip_new = int(zip_original[-5:])

        in_file.at[index, 'ZIP'] = zip_new

get_zip(popu)
popu

result_df = crime_counts.merge(popu, on='ZIP', how='left')
result_df = result_df[['ZIP', 'Crime_Type', 'Crime_Count', 'Population']]
result_df['Population'] = result_df['Population'].fillna(0)
result_df

Calculate crime rate by ZIP codes

def calculate_rate(row):
    if row['Population'] > 0:
        rate = row['Crime_Count'] / row['Population']
    else:
        rate = 1
    
    return rate

result_df['Crime_Rate'] = result_df.apply(calculate_rate, axis=1)
result_df

output_directory = 'data/'

if not os.path.exists(output_directory):
    os.makedirs(output_directory)
    
result_df.to_csv('data/crime_rate_2020.csv')
