==> 04_airbnb_basics.ipynb <==
# Obtain Information from Airbnb Listings Data
2023-10-22<br>
Evangeline Chang

import pandas as pd
import re

import warnings
warnings.filterwarnings("ignore")

listings_d = pd.read_csv('data/detailed_listings_test.csv')
print(listings_d.shape)

# see if the host is a superhost
def superhost(host):
    if host == 't':
        return 'Yes'
    else:
        return 'No'
    
listings_d['SuperHost'] = listings_d['host_is_superhost'].apply(superhost)

# get basic information of each Airbnb listing
def basic_info(in_file):
    for index, row in in_file.iterrows():
        text = row['name']

        if 'bedroom' not in text:
            if 'Studio' in text:
                in_file.at[index, 'bedrooms_n'] = 'Studio'
            else:
                in_file.at[index, 'bedrooms_n'] = 'Unknown'
        if 'bed' not in text and 'bedroom' not in text:
            in_file.at[index, 'beds_n'] = 'Unknown'
        if 'bed' not in text and 'bedroom' in text:
            in_file.at[index, 'beds_n'] = 'Unknown'
        if 'bath' not in text:
            in_file.at[index, 'baths_n'] = 'Unknown'

        split = text.split(' · ')
        for s in split:
            if 'bedroom' in s:
                bedroom = int(re.findall(r'[0-9*]', s)[0])
                in_file.at[index, 'bedrooms_n'] = bedroom
            elif 'bed' in s:
                bed = int(re.findall(r'[0-9*]', s)[0])
                in_file.at[index, 'beds_n'] = bed
            elif 'bath' in s:
                if 'Half-bath' in s or 'half-bath' in s:
                    bath = 0.5
                else:
                    bath = int(re.findall(r'[0-9*]', s)[0])
                in_file.at[index, 'baths_n'] = bath

basic_info(listings_d)

# exclude the columns that are not needed
listings_d_new = listings_d[['id', 'listing_url', 'name', 'host_id', 'latitude', 'longitude', 
                             'property_type', 'room_type', 'price', 'accommodates', 
                             'review_scores_rating', 'review_scores_accuracy', 'review_scores_cleanliness', 
                             'review_scores_checkin', 'review_scores_communication', 'review_scores_location', 'review_scores_value', 
                             'SuperHost', 'bedrooms_n', 'beds_n', 'baths_n']]

listings_d_new.to_csv('data/Airbnb_BasicInfo.csv')
