==> 06_keyword_extraction.ipynb <==
# Performed Keyword Extraction on Reviews
2023-11-03<br>
Evangeline Chang

# %pip install yake

import yake
import pandas as pd
from tqdm import tqdm
import re
import string

import warnings
warnings.filterwarnings("ignore")

review_after2020 = pd.read_csv('data/reviews_2020_check.csv', index_col=0)

review_after2020['comments'] = review_after2020['comments'].str.strip()
review_after2020 = review_after2020[review_after2020['comments'] != '']

# drop na
review_after2020.dropna(inplace=True)

# remove the rows with less than 5 characters
mask = review_after2020['comments'].str.len() >= 5
review_after2020 = review_after2020[mask]

review_after2020.reset_index(drop=True, inplace=True)
len(review_after2020)

# random non-english words from languages that also uses the latin script
# remove those rows not written in English
non_english_words = [re.compile(r'\bestancia\b', re.IGNORECASE), re.compile(r'\bzgodne\b', re.IGNORECASE),
    re.compile(r'\buma\b', re.IGNORECASE), re.compile(r'\bEspao\b', re.IGNORECASE), re.compile(r'\bett\b', re.IGNORECASE),
    re.compile(r'\boleskelu\b', re.IGNORECASE), re.compile(r'\bmuy\b', re.IGNORECASE), re.compile(r'\bflott\b', re.IGNORECASE),
    re.compile(r'\bbairro\b', re.IGNORECASE), re.compile(r'\bOrt\b', re.IGNORECASE), re.compile(r'\bpiacere\b', re.IGNORECASE),
    re.compile(r'\banfitrión\b', re.IGNORECASE), re.compile(r'\berittäin\b', re.IGNORECASE), re.compile(r'\blugar\b', re.IGNORECASE),
    re.compile(r'\bendroit\b', re.IGNORECASE), re.compile(r'\bHaus\b', re.IGNORECASE), 
    re.compile(r'\bEstadia\b', re.IGNORECASE), re.compile(r'\bvecindario\b', re.IGNORECASE), re.compile(r'\bGroßartig\b', re.IGNORECASE),
    re.compile(r'\bEstuvimos\b', re.IGNORECASE), re.compile(r'\bisäntä\b', re.IGNORECASE),  re.compile(r'\bdet\b', re.IGNORECASE),
    re.compile(r'\bEstdio\b', re.IGNORECASE), re.compile(r'\bmycket\b', re.IGNORECASE), re.compile(r'\bum\b', re.IGNORECASE),
    re.compile(r'\bestaba\b', re.IGNORECASE), re.compile(r'\blocalização\b', re.IGNORECASE), re.compile(r'\bes\b', re.IGNORECASE),
    re.compile(r'\bophold\b', re.IGNORECASE), re.compile(r'\bamfitriona\b', re.IGNORECASE), re.compile(r'\bgeweldig\b', re.IGNORECASE),
    re.compile(r'\bein\b', re.IGNORECASE), re.compile(r'\bduda\b', re.IGNORECASE), re.compile(r'\buna\b', re.IGNORECASE),
    re.compile(r'\blimpio\b', re.IGNORECASE), re.compile(r'\bbueno\b', re.IGNORECASE), re.compile(r'\bgastheer\b', re.IGNORECASE),
    re.compile(r'\bsted\b', re.IGNORECASE), re.compile(r'\bottimo\b', re.IGNORECASE), re.compile(r'\bgastvrouw\b', re.IGNORECASE),
    re.compile(r'\bposizione\b', re.IGNORECASE), re.compile(r'\bet\b', re.IGNORECASE), re.compile(r'\bnaapurusto\b', re.IGNORECASE),
    re.compile(r'\bséjour\b', re.IGNORECASE), re.compile(r'\baimer\b', re.IGNORECASE), re.compile(r'\bgénial\b', re.IGNORECASE),
    re.compile(r'\bgrozav\b', re.IGNORECASE), re.compile(r'\bfoarte\b', re.IGNORECASE), re.compile(r'\bvert\b', re.IGNORECASE),
    re.compile(r'\btycka\b', re.IGNORECASE), re.compile(r'\bestadia\b', re.IGNORECASE), re.compile(r'\been\b', re.IGNORECASE),
    re.compile(r'\bamfitrió\b', re.IGNORECASE), re.compile(r'\bgazdă\b', re.IGNORECASE), re.compile(r'\bEst\b', re.IGNORECASE),
    re.compile(r'\bfoi\b', re.IGNORECASE), re.compile(r'\bEspacio\b', re.IGNORECASE), re.compile(r'\blide\b', re.IGNORECASE),
    re.compile(r'\bbuen\b', re.IGNORECASE), re.compile(r'\bmahtava\b', re.IGNORECASE), re.compile(r'\bplăcea\b', re.IGNORECASE),
    re.compile(r'\bnabolag\b', re.IGNORECASE), re.compile(r'\bAufenthalt\b', re.IGNORECASE), re.compile(r'\bden\b', re.IGNORECASE),
    re.compile(r'\bplacering\b', re.IGNORECASE), re.compile(r'\btrès\b', re.IGNORECASE), re.compile(r'\bel\b', re.IGNORECASE),
    re.compile(r'\bubicación\b', re.IGNORECASE), re.compile(r'\bluogo\b', re.IGNORECASE), re.compile(r'\bquartier\b', re.IGNORECASE),
    re.compile(r'\bloc\b', re.IGNORECASE), re.compile(r'\bvärd\b', re.IGNORECASE), re.compile(r'\bLage\b', re.IGNORECASE),
    re.compile(r'\bhyvä\b', re.IGNORECASE), re.compile(r'\bgracias\b', re.IGNORECASE), re.compile(r'\bsoggiorno\b', re.IGNORECASE),
    re.compile(r'\bmolto\b', re.IGNORECASE), re.compile(r'\bbarri\b', re.IGNORECASE), re.compile(r'\bverblijf\b', re.IGNORECASE),
    re.compile(r'\bplaats\b', re.IGNORECASE), re.compile(r'\bEstuvo\b', re.IGNORECASE), re.compile(r'\bagradar\b', re.IGNORECASE),
    re.compile(r'\bótimo\b', re.IGNORECASE), re.compile(r'\bestada\b', re.IGNORECASE), re.compile(r'\bgrannskap\b', re.IGNORECASE),
    re.compile(r'\bsijainti\b', re.IGNORECASE), re.compile(r'\bmeget\b', re.IGNORECASE), re.compile(r'\banfitrião\b', re.IGNORECASE),
    re.compile(r'\bGastgeber\b', re.IGNORECASE), re.compile(r'\bmaison\b', re.IGNORECASE), re.compile(r'\bfantastisk\b', re.IGNORECASE),
    re.compile(r'\bkan\b', re.IGNORECASE), re.compile(r'\bospite\b', re.IGNORECASE), re.compile(r'\blocație\b', re.IGNORECASE),
    re.compile(r'\bveldig\b', re.IGNORECASE), re.compile(r'\bom\b', re.IGNORECASE), re.compile(r'\bmuito\b', re.IGNORECASE),
    re.compile(r'\bbé\b', re.IGNORECASE), re.compile(r'\blloc\b', re.IGNORECASE), re.compile(r'\bleuk\b', re.IGNORECASE),
    re.compile(r'\bsehr\b', re.IGNORECASE), re.compile(r'\bbian\b', re.IGNORECASE), re.compile(r'\bhet\b', re.IGNORECASE),
    re.compile(r'\ble\b', re.IGNORECASE), re.compile(r'\berg\b', re.IGNORECASE), re.compile(r'\bder\b', re.IGNORECASE),
    re.compile(r'\bpitää\b', re.IGNORECASE), re.compile(r'\bo\b', re.IGNORECASE), re.compile(r'\bEstancia\b', re.IGNORECASE),
    re.compile(r'\bmolt\b', re.IGNORECASE), re.compile(r'\bcea\b', re.IGNORECASE), re.compile(r'\bhuis\b', re.IGNORECASE),
    re.compile(r'\bund\b', re.IGNORECASE), re.compile(r'\bplass\b', re.IGNORECASE), re.compile(r'\bEsta\b', re.IGNORECASE),
    re.compile(r'\bopphold\b', re.IGNORECASE), re.compile(r'\bmit\b', re.IGNORECASE), re.compile(r'\bGut\b', re.IGNORECASE),
    re.compile(r'\bgoed\b', re.IGNORECASE), re.compile(r'\bubicació\b', re.IGNORECASE), re.compile(r'\beine\b', re.IGNORECASE),
    re.compile(r'\bbuono\b', re.IGNORECASE), re.compile(r'\bNachbarschaft\b', re.IGNORECASE), re.compile(r'\bcasă\b', re.IGNORECASE),
    re.compile(r'\bMögen\b', re.IGNORECASE), re.compile(r'\bil\b', re.IGNORECASE), re.compile(r'\bbon\b', re.IGNORECASE),
    re.compile(r'\bgostar\b', re.IGNORECASE), re.compile(r'\bemäntä\b', re.IGNORECASE),
    re.compile(r'\bbra\b', re.IGNORECASE), re.compile(r'\bdas\b', re.IGNORECASE), re.compile(r'\blocatie\b', re.IGNORECASE),
    re.compile(r'\bhôte\b', re.IGNORECASE), re.compile(r'\bvinden\b', re.IGNORECASE), re.compile(r'\bEste\b', re.IGNORECASE),
    re.compile(r'\bemplacement\b', re.IGNORECASE), re.compile(r'\bgenial\b', re.IGNORECASE), re.compile(r'\bvært\b', re.IGNORECASE),
    re.compile(r'\bquartiere\b', re.IGNORECASE), re.compile(r'\bde\b', re.IGNORECASE), re.compile(r'\bbom\b', re.IGNORECASE),
    re.compile(r'\bune\b', re.IGNORECASE), re.compile(r'\bședere\b', re.IGNORECASE), re.compile(r'\bplats\b', re.IGNORECASE),
    re.compile(r'\bvistelse\b', re.IGNORECASE), re.compile(r'\bpaikka\b', re.IGNORECASE), re.compile(r'\bgustar\b', re.IGNORECASE),
    re.compile(r'\bbuurt\b', re.IGNORECASE), re.compile(r'\btalo\b', re.IGNORECASE), re.compile(r'\bhus\b', re.IGNORECASE)]

def contains_non_english(text):
    for pattern in non_english_words:
        if pattern.search(text):
            return True
    return False

tqdm.pandas()
review_after2020['Contains_Non_English'] = review_after2020['comments'].progress_apply(contains_non_english)
review_after2020 = review_after2020[~review_after2020['Contains_Non_English']]
review_after2020.drop(columns=['Contains_Non_English'], inplace=True)
# review_after2020_eng = review_after2020
len(review_after2020)

# remove <br/> and apostrophes
def empty_lines_apostrophes(text):
    text = text.replace('<br/>', '')
    text = text.replace("‘", "'").replace("’", "'").replace("'s", '').replace("'ll", ' will').replace("'m", ' am').replace("'re", ' are').replace("n't", ' not').replace("'ve", ' have').replace("Ive", 'I have')
    text = text.replace("you'd", 'you would').replace('youll', 'you will').replace('youre', 'you are').replace('youd', 'you would').replace('Youll', 'You will').replace('Youre', 'You are').replace('Youd', 'You would')
    return text

tqdm.pandas()
review_after2020['comments'] = review_after2020['comments'].progress_apply(empty_lines_apostrophes)

# remove non-English characters from comments
def english_only(text):
    text = str(text)
    text = text.replace('&', 'and')
    text = re.sub(r'[^0-9A-Za-z;,:.?! ]+', '', text)
    return text

tqdm.pandas()
review_after2020['comments'] = review_after2020['comments'].progress_apply(english_only)

# replace two or more consecutive punctuations
def remove_multiple_punc(text):
    return re.sub(r'[;,:.?! ]{3,}', '.', text)

tqdm.pandas()
review_after2020['comments'] = review_after2020['comments'].progress_apply(remove_multiple_punc)

# if first character is a punctuation, remove it
def remove_punctuation_if_first_char_punctuation(text):
    if text and text[0] in string.punctuation:
        return text[1:].strip(string.punctuation)
    else:
        return text

tqdm.pandas()
review_after2020['comments'] = review_after2020['comments'].progress_apply(remove_punctuation_if_first_char_punctuation)

# remove the rows if there are less than 20 letters if punctuations were removed
def count_english_letters(text):
    english_letters = re.findall(r'[a-zA-Z]', text)
    return len(english_letters)

tqdm.pandas()
review_after2020 = review_after2020[review_after2020['comments'].progress_apply(lambda x: count_english_letters(x)) >= 20]
len(review_after2020)

# add space after punctuation (if none) and remove extra spaces
def spaces_before_after_punctuation(text):
    text = re.sub(r'\s+([.,:;!?])', r'\1', text)
    text = re.sub(r'(?<=[.,;:!?])(?![ ])', ' ', text)
    return text

tqdm.pandas()
review_after2020['comments'] = review_after2020['comments'].progress_apply(spaces_before_after_punctuation)

# add period to the end of each comment (if none) in order to join them together
def add_period_to_comments(text):
    text = text.strip()
    if text[-1] not in string.punctuation:
        return text + '.'
    else:
        return text

tqdm.pandas()
review_after2020['comments'] = review_after2020['comments'].progress_apply(add_period_to_comments)

# remove unusable rows
review_after2020['comments'] = review_after2020['comments'].str.strip()
review_after2020 = review_after2020[review_after2020['comments'] != '']

mask = review_after2020['comments'].str.len() >= 5
review_after2020 = review_after2020[mask]

review_after2020 = review_after2020[review_after2020['comments'].str.contains(' ')]

review_after2020.reset_index(drop=True, inplace=True)

len(review_after2020)

reviews_count = review_after2020.groupby(['listing_id']).size().reset_index(name='review_count').sort_values(by=['review_count'], ascending=False)
reviews_count

# filter those listings with more than 15 reviews
# After trying so many times, I think it needs at least 15 reviews for the model to run well
review_counts_15 = reviews_count[reviews_count['review_count'] >= 15]
print(len(review_counts_15))
count_15_ids = review_counts_15['listing_id'].unique()
review_after2020_15 = review_after2020[review_after2020['listing_id'].isin(count_15_ids)]

# first join commenets then run the keyword extraction model
grouped = review_after2020_15.groupby('listing_id')['comments'].apply(lambda x: " ".join(x)).reset_index()
grouped['comments'] = grouped['comments'].apply(lambda x: x.lower())

grouped['keywords_ori'] = None

kw_extractor = yake.KeywordExtractor()

tqdm.pandas()
for index, row in tqdm(grouped.iterrows(), total=len(grouped), desc="Processing rows"):
    text = row['comments']

    keywords = kw_extractor.extract_keywords(text)
    keyword_list = [word_tuple[0] for word_tuple in keywords]
    grouped.at[index, 'keywords_ori'] = keyword_list

# set words that contains little information, will remove them in the next cell

# the whole phrase will be removed if contains these words
more = ['Airbnb', 'airbnb',  'bnb', 'BNB', 'B&B', 'b&b', 'abandb', 'air bnb', 'ass', 'who', 'what', 'lol', 
        'buy', 'sooo', 'years in socal', 'dope', '20-30 minutes', 'write another glowing', 'need-a full size', 
        'boyfriend flight', 'brothers stayed', 'couple of hiccups', 'blisko', '5-10 minutes', 'minutes', 'due', 
        'drop my', 'drop off', 'los', 'tagthe']

# these words alone basically have no meaning, they can only exist with other words
# great location (OK), location (NOT OK)
single = ['site', 'website','wife', 'husband', 'absolutely', 'really', 'proximity', 'probably', 'definitely',
          'angeles', 'alley', 'back', 'night', 'side', 'host', 'time', 'people','street', 'unit', 'gonna', 'need', 
          'boyfriend', 'girlfriend', 'also', 'felt', 'listing', 'listings', 'listed', 'advertised', 'another', 'daughter',
          'away', 'coming', 'person', 'la', 'los', 'made', 'brand', 'could', 'wooow', 'didnt', 'staff', 'yortoise', 
          'guesthouse', 'guest', 'views', 'home', 'boat', 'neighborhood', 'needed', 'inn', 'hotel', 'address',
          'stay', 'place', 'would', 'area', 'spot', 'location', 'space', 'apartment', 'house', 'room','ocean', 
          'boyfriend', 'brother', 'lot', 'lots', 'lynsey', 'madre', 'make', 'year', 'rooms']

# some names that I gather from all places, needed to be removed
common_name = pd.read_csv('data/more_names.csv')
common_name['Names'] = common_name['Names'].str.strip()
common_names = list(set(common_name['Names'].tolist()))
common_names = [name.lower() for name in common_names]

more_names = ['zhenya', 'moira', 'gwen', 'abbot', 'kinney', 'zevarra', 'wolford', 'chia', 'chia mei',
              'ingy', 'iganacia', 'sora', 'iry', 'havette', 'heath', 'amberr', 'rochelle', 'nellie', 'evy', 'eryn']

# host names, needed to be removed
def abc_only(text):
    text = str(text)
    text = text.replace('&', 'and')
    text = re.sub(r'[^A-Za-z ]+', '', text)
    return text
listings = pd.read_csv('data/listings_check.csv', index_col=0)
listings['host_name'] = listings['host_name'].apply(abc_only)
host_name = listings['host_name'].tolist()
host_name = [name.strip().lower() for name in host_name if isinstance(name, str) and len(name) > 2]
host_name = list(set(host_name))

host_names = host_name
for item in host_name:
    if ' ' in item:
        split_words = item.split(' ')
        for split in split_words:
            if (len(split) > 2) and (split != 'and'):
                host_names.append(split)
host_names = [word for word in host_names if (word != 'and') and (word != '') and (len(str(word)) > 2) and (word != 'beach')]
host_names = list(set(host_names))

all_names = list(set(common_names + host_names))
all_names.extend(more_names)
all_names_s = [name + 's' if not name.endswith('s') else name for name in all_names]
all_names = all_names + all_names_s
# added space so that phrases like 'wendy's cottage' can be removed
all_names_space = [name + ' ' for name in all_names]

# remove meaningless words / phrases
grouped['keywords_filtered'] = None

tqdm.pandas()
for index, row in tqdm(grouped.iterrows(), total=len(grouped), desc="Processing rows"):
    keyword_list = row['keywords_ori']
    filtered_keyword = []
    
    for word in keyword_list:
        if (
            (word not in single) and 
            (word not in all_names) and 
            (word not in more) and 
            (all(text not in word for text in more)) and
            (all(text not in word for text in all_names_space)) and
            ('s place' not in word) and ('s home' not in word) and 
            ('. ' not in word)
        ):
            filtered_keyword.append(word)
            grouped.at[index, 'keywords_filtered'] = filtered_keyword

# get the top seven keywords of each listing
def top7(list):
    if list != None:
        return list[0:7]

grouped['top7'] = grouped['keywords_filtered'].apply(top7)

grouped['one'] = grouped['top7'].apply(lambda x: x[0] if x else None)
grouped['two'] = grouped['top7'].apply(lambda x: x[1] if x != None and len(x) > 1 else None)
grouped['three'] = grouped['top7'].apply(lambda x: x[2] if x != None and len(x) > 2 else None)
grouped['four'] = grouped['top7'].apply(lambda x: x[3] if x != None and len(x) > 3 else None)
grouped['five'] = grouped['top7'].apply(lambda x: x[4] if x != None and len(x) > 4 else None)
grouped['six'] = grouped['top7'].apply(lambda x: x[5] if x != None and len(x) > 5 else None)
grouped['seven'] = grouped['top7'].apply(lambda x: x[6] if x != None and len(x) > 6 else None)

grouped['one'] = grouped['one'].apply(lambda x: x + ', ' if x != None else None)
grouped['two'] = grouped['two'].apply(lambda x: x + ', ' if x != None else None)
grouped['three'] = grouped['three'].apply(lambda x: x + ', ' if x != None else None)
grouped['four'] = grouped['four'].apply(lambda x: x + ', ' if x != None else None)
grouped['five'] = grouped['five'].apply(lambda x: x + ', ' if x != None else None)
grouped['six'] = grouped['six'].apply(lambda x: x + ', ' if x != None else None)

# add another column; if safe is one of the keywords but not in top7, mark it
grouped['safe'] = None
grouped['safe'] = grouped.apply(lambda row: ', safe' if 'safe' in row['keywords_filtered'] and 'safe' not in row['top7'] else None, axis=1)

grouped = grouped.drop('top7', axis=1)

# this column will be used on visualization
grouped['safe_all'] = None
grouped['safe_all'] = grouped.apply(lambda row: 'Safe' if 'safe' in row['keywords_filtered'] else None, axis=1)
print(len(grouped[grouped['safe_all'] == 'Safe']))

grouped.to_csv('data/keywords.csv')

# join keywords df with the original df
listing_keywords_joined = pd.merge(listings, grouped, left_on='id', right_on='listing_id', how='left')

# this column will be used on the tooltip of the map
listing_keywords_joined['keywords_title'] = ''
listing_keywords_joined.loc[listing_keywords_joined['keywords_filtered'].notna(), 'keywords_title'] = 'Top keywords of Comments: '

# remove rows with accommodates == 0
listing_keywords_joined = listing_keywords_joined[listing_keywords_joined['accommodates'] != 0]

# fill NaN reviews with 0
columns_to_replace = [col for col in listing_keywords_joined.columns if col.startswith('review')]
listing_keywords_joined[columns_to_replace] = listing_keywords_joined[columns_to_replace].fillna(0)

listing_keywords_joined.drop(columns=['listing_id', 'comments', 'keywords_ori', 'keywords_filtered'], inplace=True)

listing_keywords_joined.to_csv('data/Airbnb_keywords.csv')
listing_keywords_joined
