==> 07_safety_related.ipynb <==
# Identify Safety-Related Reviews and Determine Unsafe Listings
2023-11-03<br>
Evangeline Chang

import pandas as pd
from tqdm import tqdm
import re
import regex

import warnings
warnings.filterwarnings("ignore")

review_after2020 = pd.read_csv('data/reviews_2020_check_safe.csv', index_col=0)
keywords = pd.read_csv('data/keywords.csv')

safe_df = review_after2020.groupby('listing_id')['comments'].apply(lambda x: " ".join(x)).reset_index()

safety_keywords = ["safe", "security", "danger", "unsafe", 'safety', 'dangerous']

# see if the comments contain safety-related content
def is_safety_related(review_text):
    for keyword in safety_keywords:
        if re.search(keyword, review_text, flags=re.IGNORECASE):
            return True
    return False

tqdm.pandas()
safe_df['Safety_Related'] = safe_df['comments'].progress_apply(is_safety_related)

# exclude rows that have gone through keyword extraction
safe_df = safe_df[safe_df['Safety_Related'] == True]

listing_ids_to_remove = keywords['listing_id'].unique()

safe_df = safe_df[~safe_df['listing_id'].isin(listing_ids_to_remove)]
safe_df.drop(columns=['Safety_Related'], inplace=True)
print(len(safe_df))

# extract the words around the keyword
def extract_context(text):
    words = re.findall(r'\b\w+\b', text)
    context_lists = []

    for i, word in enumerate(words):
        if word.lower() in safety_keywords:
            start = max(0, i - 3)
            end = min(i + 4, len(words))
            context = " ".join(words[start:end])
            context_lists.append(context)

    return context_lists

safe_df['Safety_Context'] = safe_df['comments'].apply(extract_context)

# determine whether it's unsafe or not
safe_df['not_safe_all'] = None

for index, row in safe_df.iterrows():
    context = row['Safety_Context']
    if len(context) != 0:
        if len(context) == 1:
            if 'safe' in context[0].lower() and 'not' in context[0].lower():
                safe_df.at[index, 'not_safe_all'] = 'Not Safe'
            elif 'unsafe' in context[0].lower() and 'never' not in context[0].lower() and 'not' not in context[0].lower():
                safe_df.at[index, 'not_safe_all'] = 'Not Safe'
            elif 'dangerous' in context[0].lower() and 'never' not in context[0].lower() and 'not' not in context[0].lower():
                safe_df.at[index, 'not_safe_all'] = 'Not Safe'
            
        else:
            all_items_contain_not_safe = all(['safe' in item.lower() and 'not' in item.lower() for item in context])
            all_unsafe = all(['unsafe' in item.lower() and 'never' not in item.lower() and 'not' not in item.lower() for item in context])
            if all_items_contain_not_safe:
                safe_df.at[index, 'not_safe_all'] = 'Not Safe'
            elif all_unsafe:
                safe_df.at[index, 'not_safe_all'] = 'Not Safe'

# join back to the listing file
listings = pd.read_csv('data/Airbnb_keywords.csv', index_col=0)
listing_safety_joined = pd.merge(listings, safe_df[['listing_id', 'not_safe_all']], left_on='id', right_on='listing_id', how='left')
listing_safety_joined.drop(columns=['listing_id'], inplace=True)

# join two safe-related column, will be used on visualization
for index, row in listing_safety_joined.iterrows():
    nots = row['not_safe_all']
    if nots == 'Not Safe':
        listing_safety_joined.at[index, 'safe_all'] = 'Not Safe'
    elif nots == 'Safe':
        listing_safety_joined.at[index, 'safe_all'] = 'Safe'

listing_safety_joined.drop(columns=['not_safe_all'], inplace=True)

listing_safety_joined.to_csv('data/Airbnb_final.csv')
listing_safety_joined
