==> 08_crime_bar.ipynb <==
# Categorize Crime Types and Group them by Quarters and ZIP Codes
2023-11-06<br>
Evangeline Chang

import pandas as pd
from tqdm import tqdm
from datetime import datetime

import warnings
warnings.filterwarnings("ignore", category=FutureWarning)

crime_full = pd.read_csv('data/crime_data_2020_zipcode_final.csv', index_col=0)
crime_full.head(1)

crime_full['Crime Type'] = None

tqdm.pandas()
for index, row in tqdm(crime_full.iterrows(), total=len(crime_full), desc="Processing rows"):
    type = row['Crm Cd Desc']

    if 'RAPE' in type or 'ORAL' in type or 'SEXUAL' in type or 'LEWD' in type or 'SEX,UNLAWFUL' in type or 'CHILD PORNOGRAPHY' in type:
        crime_full.at[index, 'Crime Type'] = 'Sex-Related'

    elif 'ATTEMPTED' in type or 'ATTEMPT' in type:
        crime_full.at[index, 'Crime Type'] = 'Others'
        
    elif 'ATTEMPTED' not in type:
        if 'ASSAULT' in type or 'Battery' in type:
            crime_full.at[index, 'Crime Type'] = 'Assault'
        
        elif 'THEFT' in type or 'STOLEN' in type or 'PURSE SNATCHING' in type or 'TILL TAP' in type or 'DRUNK ROLL' in type:
            crime_full.at[index, 'Crime Type'] = 'Theft'
        
        elif 'VANDALISM' in type:
            crime_full.at[index, 'Crime Type'] = 'Vandalism'
        
        elif 'BURGLARY' in type:
            crime_full.at[index, 'Crime Type'] = 'Burglary'
        
        else:
            crime_full.at[index, 'Crime Type'] = 'Others'

def get_season(in_file):
    for index, row in in_file.iterrows():
        date_str = row['DATE OCC']
        date_obj = datetime.strptime(date_str, "%m/%d/%Y %I:%M:%S %p")
        month = date_obj.month
        year = date_obj.year
        if month >= 1 and month <= 3:
            quarter = 'Q1'
        elif month >= 4 and month <= 6:
            quarter = 'Q2'
        elif month >= 7 and month <= 9:
            quarter = 'Q3'
        elif month >= 10:
            quarter = 'Q4'
        
        in_file.at[index, 'Season'] = str(year) + ' ' + quarter

get_season(crime_full)

crime_full = crime_full[crime_full['Season'] != '2023 Q4']
crime_full = crime_full.groupby(['ZIP', 'Crime Type', 'Season']).size().reset_index(name='Crime_Count')

crime_full.to_csv('data/crime_2020_viz.csv')
crime_full.head(3)
